refactor : Delegate Protocol 파일 분리 및 전체적인 코드 컨벤션 수정
